refactor to make Unitcell shape independent of full symmetry data
