+++ src/mol-plugin-state/transforms/representation.ts
@@ -1220,7 +1220,7 @@ const ModelUnitcell3D = PluginStateTransform.BuiltIn({
         return Task.create('Model Unit Cell', async ctx => {
             const symmetry = ModelSymmetry.Provider.get(a.data);
             if (!symmetry) return StateObject.Null;
-            const data = getUnitcellData(a.data, symmetry, params);
+            const data = getUnitcellData(a.data, symmetry.spacegroup.cell, params);
             const repr = UnitcellRepresentation({ webgl: plugin.canvas3d?.webgl, ...plugin.representation.structure.themes }, () => UnitcellParams);
             await repr.createOrUpdate(params, data).runInContext(ctx);
             return new SO.Shape.Representation3D({ repr, sourceData: data }, { label: `Unit Cell`, description: symmetry.spacegroup.name });
@@ -1231,7 +1231,7 @@ const ModelUnitcell3D = PluginStateTransform.BuiltIn({
             const symmetry = ModelSymmetry.Provider.get(a.data);
             if (!symmetry) return StateTransformer.UpdateResult.Null;
             const props = { ...b.data.repr.props, ...newParams };
-            const data = getUnitcellData(a.data, symmetry, props);
+            const data = getUnitcellData(a.data, symmetry.spacegroup.cell, props);
             await b.data.repr.createOrUpdate(props, data).runInContext(ctx);
             b.data.sourceData = data;
             return StateTransformer.UpdateResult.Updated;
